requirements.txt
Django==1.6.5
Markdown==2.4.1
Pygments==1.6
argparse==1.2.1
django-dotenv==1.2
django-extensions==1.3.7
django-filter==0.7
django-rest-swagger==0.1.14
djangorestframework==2.3.14
six==1.7.3
sqlparse==0.1.11
wsgiref==0.1.2
